[runner.rb] Allow for RAILS_COMMIT/rails bisect
This is a feature that allows a user to `git bisect` the rails repo
locally, and then specify a RAILS_COMMIT from that to send to the
travis run to determine if it is a "good/bad" commit.

Will use all of the rails direct dependencies from that specific commit.