[FIXUP][runner.rb] Use Repository for rails gems
Turns out, when using:

    git "https://github.com/rails/rails.git", :sha => "ABCD123" do
      gem "actioncable"
      gem "..."
    end

It won't generate a version when the bundler output runs into a
conflict, so it is harder to debug what we are using.

Example Error output:

    Fetching source index from https://rubygems.org/
    Resolving dependencies..........................................................................................
    Bundler could not find compatible versions for gem "actioncable":

      In Gemfile:
	actioncable

	manageiq-schema was resolved to 0.1.0, which depends on
	  rails (= 5.2.0.beta2) was resolved to 5.2.0.beta2, which depends on
	    actioncable (= 5.2.0.beta2)

(in this case, the first actioncable line doesn't provide a version it is being
resolved to when pulled in from a git source)

So instead, use the `Repository` class to handle the fetch, which should
use it as a `:path` gem instead of a `:git` gem under the hood, and
hopefully should work better with the error output.

As a side benefit, it should at least provide a faster fetch of the rails repo
instead of a full/partial clone... hopefully...