[gemspec] Remove ActiveSupport
For the most part, this isn't providing much value being here, and in
the places it was helping, it can be easily replaced by core Ruby
methods.

Removing it avoids issues when repos like `manageiq-content` run
automate code that will load in activesupport without the bundle
environment to version check.  This avoids needing to keep this repo in
lockstep with the rest of the `ManageIQ/*` to make sure multiple
versions of `ActiveSupport` are not installed.